app: change how client list is fetched
The old method was that a message would come in via postMessage that
indicates we should send a request to the client-list endpoint to
download a new list. Now, the message that comes via postMessage
contains the list, and the client-list endpoint has been repurposed to
request a new list via postMessage instead of in the web response. This
saves some back-and-forth between processes, and also some needless
conversion to and from a string.