main: always delete CefCache before starting CEF
It seems CEF can't be trusted to store valid cache between shutdowns or
system updates, which leads to infrequent reports of SKIA shader
compilation errors, the only fix for which is to delete CefCache or
GPUCache. If that's the case, I won't let it store anything at all. This
doesn't appear to affect startup time.