window_launcher: append file separator
This is required by the plugin loader, and is usually handled by the
frontend but we don't get the path from the frontend in this specific
case, so we need to append one ourselves. We also need to make sure the
path uses '/' as separators, but as far as I can tell, C++'s "path" type
will always give us that as long as we don't call "make_preferred".