library: handle more xcb_input_ events
XCB_INPUT_* events are sometimes used instead of, or as well as, the
standard XCB_* equivalents. In particular these event types are
seemingly used more in SDL3 than SDL2, but I've only been testing this
with SDL2 so far. This commit also cleans up a bunch of copy-paste code
in the xcb handlers, and adds checks to make sure XCB_INPUT_ events are
destined for the specific window we're interested in.