window_launcher: lots of deduplication
As well as fixing some inconsistent behaviour, the main purpose of this
is to make sure either exec() or exit() is always called on a fork()-ed
child process. With custom launch commands, it's now possible to cause
exec() to fail, which was resulting in the child process just falling
through and continuing alongside the parent process - which usually ends
up causing the whole program to crash.