library: always init on MakeCurrent
Normally, the game runs [e/w]glInitialize, then creates a context, binds
it, uses it for stuff, destroys it, calls [e/w]glTerminate, then calls
Initialize again but on a different thead, that being the main thread.
To work around this, I assumed that the second context to be created
(with no shared_context) will always be the main one, and run my own
initialization on the thread that context gets bound to.

But somehow, someone has managed to find a case where it doesn't do
this, and instead the first context it creates is the main one. This
might be specific to Intel GPUs on Windows - not sure.

The most robust way to fix this appears to be always initialising when a
main context is created and always terminating when it gets destroyed.
This unfortunately means that for everyone else it will init twice,
slowing down startup. But I can't think of a better solution right now.
(Maybe the game engine itself should be fixed to not do such insane
things in the first place, but that's a rant for another place.)